requirements.txt
lightgbm==3.3.3
mlflow==1.29.0
numpy==1.23.4
optuna==3.0.3
pandas==1.3.5
PyYAML==6.0
scikit-learn==1.0.2
seaborn==0.12.1
transformers==4.23.1
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.2+cu113 ; python_version < '3.9'
python-dotenv==0.21.0
scikit-multilearn==0.2.0
sentencepiece==0.1.97
# protobuf==4.21.9
